pyproject.toml
[project]
name = "blog-melcore-fr"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "jinja2==3.1.6",
    "markdown>=3.10.2",
    "pygments>=2.19.2",
    "python-slugify>=8.0.4",
    "text-unidecode>=1.3",
]
